LayoutTests/platform/mac-sequoia-wk2/imported/w3c/web-platform-tests/css/css-tables/table-model-fixup-2-expected.txt

Table fixup

Check that replaced elements display types are handled properly in fixup

Both input fields should look identical, on their own line:
Replaced elements inside a table cannot be table-row and are considered inline -- input elements (width)



Both input fields should look identical and be on the same line:
Replaced elements inside a table cannot be table-row and are considered inline -- input elements (top)


Both images should look identical, on their own line:
Replaced elements inside a table cannot be table-row and are considered inline -- img elements (width)



Both images should look identical and be on the same line:
Replaced elements inside a table cannot be table-row and are considered inline -- img elements (top)


Both input fields should look identical, on their own line:
Replaced elements inside a table cannot be table-column and are considered inline -- input elements (width)



Both input fields should look identical and be on the same line:
Replaced elements inside a table cannot be table-column and are considered inline -- input elements (top)


Both images should look identical, on their own line:
Replaced elements inside a table cannot be table-column and are considered inline -- img elements (width)



Both images should look identical and be on the same line:
Replaced elements inside a table cannot be table-column and are considered inline -- img elements (top)


Both input fields should look identical, on their own line:
Replaced elements inside a table cannot be table-cell and are considered inline -- input elements (width)



Both input fields should look identical and be on the same line:
Replaced elements inside a table cannot be table-cell and are considered inline -- input elements (top)


Both images should look identical, on their own line:
Replaced elements inside a table cannot be table-cell and are considered inline -- img elements (width)



Both images should look identical and be on the same line:
Replaced elements inside a table cannot be table-cell and are considered inline -- img elements (top)


Both text inputs should share the same line:
Replaced elements outside a table cannot be inline-table and are considered inline -- input=text elements


Both button inputs should share the same line:
Replaced elements outside a table cannot be inline-table and are considered inline -- input=button elements


Both file inputs should share the same line:
Replaced elements outside a table cannot be inline-table and are considered inline -- input=file elements


Both images should share the same line:
Replaced elements outside a table cannot be inline-table and are considered inline -- img elements


Both text inputs should stand on their own line:
Replaced elements outside a table cannot be table and are considered block -- input=text elements


Both button inputs should stand on their own line:
Replaced elements outside a table cannot be table and are considered block -- input=button elements


Both file inputs should stand on their own line:
Replaced elements outside a table cannot be table and are considered block -- input=file elements


Both images should stand on their own line:
Replaced elements outside a table cannot be table and are considered block -- img elements


Both text inputs should share the same line:
Replaced elements outside a table cannot be table-row and are considered inline -- input=text elements


Both button inputs should share the same line:
Replaced elements outside a table cannot be table-row and are considered inline -- input=button elements


Both file inputs should share the same line:
Replaced elements outside a table cannot be table-row and are considered inline -- input=file elements


Both images should share the same line:
Replaced elements outside a table cannot be table-row and are considered inline -- img elements


Both text inputs should share the same line:
Replaced elements outside a table cannot be table-row-group and are considered inline -- input=text elements


Both button inputs should share the same line:
Replaced elements outside a table cannot be table-row-group and are considered inline -- input=button elements


Both file inputs should share the same line:
Replaced elements outside a table cannot be table-row-group and are considered inline -- input=file elements


Both images should share the same line:
Replaced elements outside a table cannot be table-row-group and are considered inline -- img elements


Both text inputs should share the same line:
Replaced elements outside a table cannot be table-column and are considered inline inline -- input=text elements


Both button inputs should share the same line:
Replaced elements outside a table cannot be table-column and are considered inline -- input=button elements


Both file inputs should share the same line:
Replaced elements outside a table cannot be table-column and are considered inline -- input=file elements


Both images should share the same line:
Replaced elements outside a table cannot be table-column and are considered inline -- img elements


Both text inputs should share the same line:
Replaced elements outside a table cannot be table-cell and are considered inline -- input=text elements


Both button inputs should share the same line:
Replaced elements outside a table cannot be table-cell and are considered inline -- input=button elements


Both file inputs should share the same line:
Replaced elements outside a table cannot be table-cell and are considered inline -- input=file elements


Both images should share the same line:
Replaced elements outside a table cannot be table-cell and are considered inline -- img elements


Both text inputs should share the same line:
Replaced elements outside a table cannot be table-caption and are considered inline -- input=text elements


Both button inputs should share the same line:
Replaced elements outside a table cannot be table-caption and are considered inline -- input=button elements


Both file inputs should share the same line:
Replaced elements outside a table cannot be table-caption and are considered inline -- input=file elements


Both images should share the same line:
Replaced elements outside a table cannot be table-caption and are considered inline -- img elements



PASS Replaced elements inside a table cannot be table-row and are considered inline -- input elements (width)
PASS Replaced elements inside a table cannot be table-row and are considered inline -- input elements (top)
PASS Replaced elements inside a table cannot be table-row and are considered inline -- img elements (width)
FAIL Replaced elements inside a table cannot be table-row and are considered inline -- img elements (top) assert_equals: expected 803.5625 but got 835.5625
PASS Replaced elements inside a table cannot be table-column and are considered inline -- input elements (width)
PASS Replaced elements inside a table cannot be table-column and are considered inline -- input elements (top)
PASS Replaced elements inside a table cannot be table-column and are considered inline -- img elements (width)
FAIL Replaced elements inside a table cannot be table-column and are considered inline -- img elements (top) assert_equals: expected 1551.59375 but got 1583.59375
PASS Replaced elements inside a table cannot be table-cell and are considered inline -- input elements (width)
PASS Replaced elements inside a table cannot be table-cell and are considered inline -- input elements (top)
PASS Replaced elements inside a table cannot be table-cell and are considered inline -- img elements (width)
FAIL Replaced elements inside a table cannot be table-cell and are considered inline -- img elements (top) assert_equals: expected 2289.625 but got 2321.625
PASS Replaced elements outside a table cannot be inline-table and are considered inline -- input=text elements
PASS Replaced elements outside a table cannot be inline-table and are considered inline -- input=button elements
PASS Replaced elements outside a table cannot be inline-table and are considered inline -- input=file elements
PASS Replaced elements outside a table cannot be inline-table and are considered inline -- img elements
PASS Replaced elements outside a table cannot be table and are considered block -- input=text elements
PASS Replaced elements outside a table cannot be table and are considered block -- input=button elements
PASS Replaced elements outside a table cannot be table and are considered block -- input=file elements
PASS Replaced elements outside a table cannot be table and are considered block -- img elements
PASS Replaced elements outside a table cannot be table-row and are considered inline -- input=text elements
PASS Replaced elements outside a table cannot be table-row and are considered inline -- input=button elements
FAIL Replaced elements outside a table cannot be table-row and are considered inline -- input=file elements assert_equals: expected 3725.65625 but got 3743.65625
FAIL Replaced elements outside a table cannot be table-row and are considered inline -- img elements assert_equals: expected 3857.65625 but got 3889.65625
PASS Replaced elements outside a table cannot be table-row-group and are considered inline -- input=text elements
PASS Replaced elements outside a table cannot be table-row-group and are considered inline -- input=button elements
FAIL Replaced elements outside a table cannot be table-row-group and are considered inline -- input=file elements assert_equals: expected 4248 but got 4266
FAIL Replaced elements outside a table cannot be table-row-group and are considered inline -- img elements assert_equals: expected 4380 but got 4412
PASS Replaced elements outside a table cannot be table-column and are considered inline inline -- input=text elements
PASS Replaced elements outside a table cannot be table-column and are considered inline -- input=button elements
FAIL Replaced elements outside a table cannot be table-column and are considered inline -- input=file elements assert_equals: expected 4770.34375 but got 4788.34375
FAIL Replaced elements outside a table cannot be table-column and are considered inline -- img elements assert_equals: expected 4902.34375 but got 4934.34375
PASS Replaced elements outside a table cannot be table-cell and are considered inline -- input=text elements
PASS Replaced elements outside a table cannot be table-cell and are considered inline -- input=button elements
FAIL Replaced elements outside a table cannot be table-cell and are considered inline -- input=file elements assert_equals: expected 5292.6875 but got 5310.6875
FAIL Replaced elements outside a table cannot be table-cell and are considered inline -- img elements assert_equals: expected 5424.6875 but got 5456.6875
PASS Replaced elements outside a table cannot be table-caption and are considered inline -- input=text elements
PASS Replaced elements outside a table cannot be table-caption and are considered inline -- input=button elements
FAIL Replaced elements outside a table cannot be table-caption and are considered inline -- input=file elements assert_equals: expected 5815.03125 but got 5833.03125
FAIL Replaced elements outside a table cannot be table-caption and are considered inline -- img elements assert_equals: expected 5947.03125 but got 5979.03125
